requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

tensorflow>=2.10.0
keras>=2.10.0

matplotlib>=3.5.0
seaborn>=0.11.0

scipy>=1.9.0

# Crawler dependencies
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0
lxml>=4.9.0
python-dateutil>=2.8.0
openpyxl>=3.1.0

# Server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
python-multipart>=0.0.6
requests>=2.31.0
joblib>=1.3.0
holidays>=0.36
PyJWT>=2.8.0
PyMySQL>=1.1.0
python-dotenv>=1.0.1

